Add date-fns, adjust template, add task modules
In order to implement all the features I am looking for, I am reaching
for date-fns to simplify some of the logic and timing concerning due
dates (specifically checking if a task is due today or overdue). I have
also initialized a Task class that has all the methods I will need for
the features that I want to implement(Like being able to edit the task
as well as mark it as completed). I have also started working on the
Project class that will essentially work as a task manager, storing
tasks and taking in the data required to add new ones. I still need to
figure out how to add a default project that can't be removed, and I
also need to read up on using localStorage to see if I need to update
my classes to enable that functionality.